requirements.txt
# Core ETL dependencies
# psycopg v3 (psycopg[binary]) is preferred; keep psycopg2-binary temporarily for compatibility
psycopg[binary]
psycopg2-binary
pandas
python-dotenv>=1.0
phonenumbers
typer
pydantic>=2.7
loguru
httpx>=0.27
tenacity>=8.0.0
pyyaml>=5.4
pytest
pytest-asyncio>=0.23.0
fastapi
uvicorn
